test(iterate): split review-summary minimize tests to satisfy 200-line oxlint cap
The rewritten seen-summary self-minimize test file grew past oxlint's
200-line max after covering issue #313. Split first-look-specific cases
into their own file; no behavioral change.
